Get the project building with Clang 3.1.
 * Clang 3.1 is not compatible with the chrono library in libstdc++ 4.6
   and 4.7.  (Apparently the chrono library has a bug.)  The bug in chrono
   makes it impossible to include mutex, which contains std::mutex and
   std::lock_guard.  Therefore, replace these classes with
   indexer::Mutex and indexer::LockGuard<T>, where indexer::Mutex uses
   pthreads on Unix (and std::mutex otherwise).

 * Harmonize the use of the class/struct keyword between forward
   declarations and class/struct definitions.  Clang warns about
   mismatches.

 * Clang complained about the discarded result of waitpid.  Instead of
   discarding it, assert that it is the PID we're waiting for.

 * After switching from GCC to Clang, the QMAKE_LINK_C variable was not
   set, breaking the hack in libbtrace/libbtrace.pro that sets QMAKE_LINK.
   Use QMAKE_CC instead.

 * After switching from GCC to Clang, I noticed that the -lpthread
   command-line argument was causing libpthread.so.0 to appear in the ldd
   output for libbtrace.so.  Suppress -lpthread by clearing
   QMAKE_LIBS_THREAD.

 * Fix a defect in libindexdb/Buffer.h.  The argument to the copy
   constructors must be const-qualified.  The Clang error was not easy to
   understand -- it complained about the copy constructor of the
   std::pair<StringTable, std::vector<ID>> class.  (Buffer and StringTable
   are both intended to be non-copyable.)

 * Remove a pointless cast in Table::lowerBound.  row.count() and
   columnCount() both have type int.  Clang warned about this.

 * Explicitly declare a defaulted StringTable::operator=(StringTable &&).
   (Apparently G++ 4.6 implicitly declares this?)

 * Clang complains that Ref is incomplete in Project::queryFileRefs<Func>,
   in Project.h.  Ref.h includes Project.h, so we can't fix this problem by
   including Ref.h at the top of Project.h.  Instead, move the template
   function into Project-inl.h, which is included in SourceWidget.cc.

+++ libindexdb/IndexDb.cc
@@ -290,7 +290,7 @@ void Table::setReadOnly(
 TableIterator Table::lowerBound(const Row &row)
 {
     assert(m_readonly);
-    assert(static_cast<size_t>(row.count()) <= columnCount());
+    assert(row.count() <= columnCount());
 
     TableIterator itMin = begin();
     TableIterator itMax = end();
